fix: placedetail header 수정
